Improve parameter block handling during update
This uses the new feature of ra-pb-create to ensure that the parameter
block has the expected version and binary layout of the running or
to-be-installed safety firmware, even when the tools already support
newer parameter block versions.

It also ensures that when downgrading, the parameter block is
downgraded, too.

+++ firmware/ra-update.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 #
 # This script checks whether the Safety Controller needs an update.
-# For this, it checks the current flashed firmware information and compares
-# it with the data from the firmware bundle in the rootfs.
+# For this, it checks the current flashed firmware information (CURRENT_...)
+# and compares it with the data from the firmware bundle in the rootfs (TARGET_...).
 # If this differs, then the firmware is updated.
 #
 
@@ -19,27 +19,14 @@ PARAM_FILE="$(ls -1 $LIBDIR/*_parameter-block_only-contactor.yaml 2>/dev/null)"
 TARGET_VERSION="$(ra-update fw-info "$FW_FILE" 2>/dev/null)"
 CURRENT_VERSION="$(ra-update fw-info)"
 
+TARGET_PB_VERSION="$(echo "$TARGET_VERSION" | awk -F': *' '/Parameter Block Version/ {print $2}')"
+CURRENT_PB_VERSION="$(echo "$CURRENT_VERSION" | awk -F': *' '/Parameter Block Version/ {print $2}')"
+
 CMP_TARGET_VERSION="$(echo "$TARGET_VERSION" | tail -n +3 | head -n 6)"
 CMP_CURRENT_VERSION="$(echo "$CURRENT_VERSION" | tail -n +3 | head -n 6)"
 
-# nothing to do when versions are equal
-[ "$CMP_TARGET_VERSION" = "$CMP_CURRENT_VERSION" ] && exit 0
-
-cat <<EOF
-== Safety Controller Firmware Update required ==
-
-<<< Current Firmware <<<
-$CMP_CURRENT_VERSION
-<<<
-
->>> New Firmware >>>
-$CMP_TARGET_VERSION
->>>
-
-================================================
-EOF
-
 CURRENT_VERSION_ONLY="$(echo "$CMP_CURRENT_VERSION" | grep "Version" | awk '{print $3}')"
+TARGET_VERSION_ONLY="$(echo "$CMP_TARGET_VERSION" | grep "Version" | awk '{print $3}')"
 
 version_cmp() {
     local v1="$1"
@@ -71,20 +58,46 @@ cur_version_gt() {
     [ "$(version_cmp "$CURRENT_VERSION_ONLY" "$1")" = "1" ]
 }
 
+# greater than or equal
+cur_version_ge() {
+    local rv="$(version_cmp "$CURRENT_VERSION_ONLY" "$1")"
+    [ "$rv" = "0" ] || [ "$rv" = "1" ]
+}
+
 # equal
 cur_version_eq() {
     [ "$CURRENT_VERSION_ONLY" = "$1" ]
 }
+
+# nothing to do when versions are equal
+[ "$CMP_TARGET_VERSION" = "$CMP_CURRENT_VERSION" ] && exit 0
+
+cat <<EOF
+== Safety Controller Firmware Update required ==
+
+<<< Current Firmware <<<
+$CMP_CURRENT_VERSION
+<<<
+
+>>> New Firmware >>>
+$CMP_TARGET_VERSION
+>>>
+
+================================================
+EOF
 
 # when upgrading from 0.1.0 we need to install a parameter block for the first time;
 # later we assume that a valid parameter block is already installed and we don't touch it
 if cur_version_eq "0.1.0"; then
     pbfile_target_bin=$(mktemp)
 
-    ra-pb-create -i "$PARAM_FILE" -o "$pbfile_target_bin"
+    ra-pb-create -O "$TARGET_PB_VERSION" -i "$PARAM_FILE" -o "$pbfile_target_bin"
 
     echo -n "Installing Parameter Block..."
     ra-update -a data flash "$pbfile_target_bin"
+
+    # remember the updated pb version
+    CURRENT_PB_VERSION="$TARGET_PB_VERSION"
 
     rm -f "$pbfile_target_bin"
     echo "done."
@@ -110,17 +123,19 @@ if cur_version_gt "0.1.0" && cur_version_le "0.2.2"; then
 
     # re-create parameter block - while creating YAML, the old magic values was already handled
     # so we can just re-create the new binary file using the dumped YAML file
-    ra-pb-create -i "$pbfile_current_yaml" -o "$pbfile_target_bin"
+    ra-pb-create -O "$TARGET_PB_VERSION" -i "$pbfile_current_yaml" -o "$pbfile_target_bin"
 
     # finally flash it
     ra-update -a data flash "$pbfile_target_bin"
+
+    # remember the updated pb version
+    CURRENT_PB_VERSION="$TARGET_PB_VERSION"
 
     rm -f "$pbfile_current_bin" "$pbfile_current_yaml" "$pbfile_target_bin"
     echo "done."
 fi
 
-# firmware < 0.2.6 had no versioned parameter file, ra-pb-dump can handle this
-# but we need to write it back
+# firmware < 0.2.6 had no versioned parameter file
 if cur_version_lt "0.2.6"; then
     echo "Updating Parameter Block..."
 
@@ -137,10 +152,42 @@ if cur_version_lt "0.2.6"; then
     cat "$pbfile_current_yaml" | sed 's/^/    /'
 
     # re-create parameter block
-    ra-pb-create -i "$pbfile_current_yaml" -o "$pbfile_target_bin"
+    ra-pb-create -O "$TARGET_PB_VERSION" -i "$pbfile_current_yaml" -o "$pbfile_target_bin"
 
     # finally flash it
     ra-update -a data flash "$pbfile_target_bin"
+
+    # remember the updated pb version
+    CURRENT_PB_VERSION="$TARGET_PB_VERSION"
+
+    rm -f "$pbfile_current_bin" "$pbfile_current_yaml" "$pbfile_target_bin"
+    echo "done."
+fi
+
+# check whether the parameter block versions differ (e.g. due to downgrade)
+if [ "$CURRENT_PB_VERSION" != "$TARGET_PB_VERSION" ]; then
+    echo "Adapting Parameter Block..."
+
+    pbfile_current_bin=$(mktemp)
+    pbfile_current_yaml=$(mktemp)
+    pbfile_target_bin=$(mktemp)
+
+    # dump current parameter block to YAML file
+    ra-update -a data dump "$pbfile_current_bin"
+    ra-pb-dump "$pbfile_current_bin" > "$pbfile_current_yaml"
+
+    # dump it indented to stdout (for debug purpose only)
+    echo "Current parameters:"
+    cat "$pbfile_current_yaml" | sed 's/^/    /'
+
+    # re-create parameter block - maybe with warnings about unsupported fields
+    ra-pb-create -O "$TARGET_PB_VERSION" -i "$pbfile_current_yaml" -o "$pbfile_target_bin"
+
+    # finally flash it
+    ra-update -a data flash "$pbfile_target_bin"
+
+    # remember the updated pb version
+    CURRENT_PB_VERSION="$TARGET_PB_VERSION"
 
     rm -f "$pbfile_current_bin" "$pbfile_current_yaml" "$pbfile_target_bin"
     echo "done."
